% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array2list.R
\name{array2list}
\alias{array2list}
\title{Reorganize an array of graphs to a list of graphs.}
\usage{
array2list(list_in)
}
\arguments{
\item{array_out}{an array of dimensions [n x m x p].}
}
\value{
list_in a list with n elements of dimensions [n x m].
}
\description{
\code{array2list} uses plyr to reshape an array of same-dimensional graphs
a list. Takes from array[,,length(listels)] to list[[listels]].
}
\examples{
test <- array(1:105, c(5, 7, 3))
array2list(test)  # has 3 elements with dimensions 5x7
}
\seealso{
\code{\link{plyr}} \code{\link{list2array}}
}
