% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_estimator.R
\name{beta_estimator}
\alias{beta_estimator}
\title{A function to fit a beta distribution to a given univariate sample of data.}
\usage{
beta_estimator(sample)
}
\arguments{
\item{sample}{a univariate sample of values with arbitrary class labels.}
}
\value{
params$alpha the alpha parameter per edge. [n x m]

params$beta the beta parameter per edge. [n x m]
}
\description{
\code{beta_estimator} uses the method of moments to estimate the parameters of a beta
distribution, alpha and beta, for a given sample.
}
\seealso{
\code{\link{beta_graph_estimator}}
}
