% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_graph_estimator.R
\name{beta_graph_estimator}
\alias{beta_graph_estimator}
\title{A function to fit a beta distribution to the edges of the matrices in a graph.}
\usage{
beta_graph_estimator(sample)
}
\arguments{
\item{object}{a list or array of graphs with arbitrary labelling.
 - if object is a list, then it should have p elements of dimensions
     [n x m].
- if object is an array, then it should be of dimensions [n x m x p].}
}
\value{
params$alpha the alpha parameter per edge. [n x m]

params$beta the beta parameter per edge. [n x m]
}
\description{
\code{beta_graph_estimator} uses the method of moments to estimate the parameters of a beta
distribution, alpha and beta, for a collection of graphs.
}
\seealso{
\code{\link{beta_graph_estimator}}
}
