% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2array.R
\name{list2array}
\alias{list2array}
\title{Reorganize a list of graphs to an array of graphs}
\usage{
list2array(list_in)
}
\arguments{
\item{list_in}{a list with n elements of dimensions [m x m].}
}
\value{
array_out an array of dimensions [m x m x n].
}
\description{
\code{list2array} uses abind to reshape a list of same-dimension graphs to
an array. Takes from list[[listels]] to array[,,length(listels)].
}
\examples{
test <- list()
test[[1]] <- matrix(1:35, nrow=5, ncol=7)
test[[2]] <- matrix(36:70, nrow=5, ncol=7)
test[[3]] <- matrix(71:105, nrow=5, ncol=7)
list2array(test)  # has dimensions [5, 7, 3]
}
\seealso{
\code{\link{abind}} \code{\link{array2list}}
}
