% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_beta_graph.R
\name{sample_beta_graph}
\alias{sample_beta_graph}
\title{Sample Graphs from Beta Distribution}
\usage{
sample_beta_graph(alpha, beta, s = 10, type = "array")
}
\arguments{
\item{alpha}{a [n x m] matrix indicating the alphas of each edge.}

\item{beta}{a [n x m] matrix indicating the betas of each edge.}

\item{type}{an option of "list" or "array" (default) for the format of the output.}

\item{n}{the number of graphs to sample.}
}
\value{
sample the graph observations from the given beta distribution.
    - if type == "list", returns a p element list of [n x m] observations.
    - if type == "array" (default), returns a [n x m x s] element array where the 3rd dimension indexes the observations.
}
\description{
\code{sample_beta_graph} A function to sample from a graph-valued RV where each edge follows a beta distribution.
}
\seealso{
\code{\link{list2array}} \code{\link{array2list}}
}
