% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_estimator.R
\name{sg.bern.graph_estimator}
\alias{sg.bern.graph_estimator}
\title{A function to fit a bernoulli distribution to the edges of the matrices in a graph.}
\usage{
sg.bern.graph_estimator(samp, thresh = 0, smooth = TRUE)
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
 - if samp is a list, then it should have s elements of dimensions
     [n x m].
- if samp is an array, then it should be of dimensions [n x m x s].}

\item{thresh=0:}{is the threshold below which we set edges to disconnected, and above which we set edges to connected.}
}
\value{
params$alpha the alpha parameter per edge. [n x m]

params$beta the beta parameter per edge. [n x m]
}
\description{
\code{sg.bern.graph_estimator} uses the method of moments to estimate the parameters of a beta
distribution, alpha and beta, for a collection of graphs.
}
\seealso{
\code{\link{sg.bern.estimator}}
}
