% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph_classifier.R
\name{sg.bern.subgraph_classifier}
\alias{sg.bern.subgraph_classifier}
\title{Bernoulli Subgraph Classifier}
\usage{
sg.bern.subgraph_classifier(test, stats, p, pi, e, classes, coherent = FALSE)
}
\arguments{
\item{test}{the test array to generate labels for.
 - if test is a list, then it should have s elements of dimensions
     [n x n].
- if test is an array, then it should be of dimensions [n x n x s].}

\item{p}{[n x n x c] the probability per edge of being connected per class for c classes.}

\item{pi}{[c] the pi vector probability of each class occuring.}

\item{e}{[1] the number of edges in the subgraph.}

\item{classes}{the labels for the classes as ordered in the pi and p arrays.}

\item{coherent=FALSE}{if FALSE, estimate an incoherent subgraph, otherwise an integer indicating the number of vertices in the coherent subgraph.}
}
\value{
edges [e] the edges present in the subgraph.

Yhat [s] the predicted class for each test example.
}
\description{
\code{sg.bern.subgraph_classifier} classifies arbitrary data given the estimators of a subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
