% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph_train.R
\name{sg.bern.subgraph_train}
\alias{sg.bern.subgraph_train}
\title{Bernoulli Subgraph Train}
\usage{
sg.bern.subgraph_train(samp, Y, tstat = "fisher")
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
 - if samp is a list, then it should have s elements of dimensions
     [n x n].
- if samp is an array, then it should be of dimensions [n x n x s].}

\item{Y}{[s] the class labels.}

\item{tstat="fisher"}{the test statistic to use. options are fisher's exact ("fisher") and chi-squared ("chisq").}
}
\value{
edges [n x n] an ordering of the edges in the subgraph by test statistic results.

p [n x n x c] the probability per edge of being connected per class for c classes.

pi [c] the probability of seeing a given class.
}
\description{
\code{sg.bern.xval_classifier} Trains a Model for identification of the edges in a subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
