% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval_subgraph_classifier.R
\name{sg.bern.xval_classifier}
\alias{sg.bern.xval_classifier}
\title{Bernoulli Subgraph Classifier}
\usage{
sg.bern.xval_classifier(samp, Y, test_stats, p, pi, es, coherent = FALSE,
  tstat = "fisher")
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
 - if samp is a list, then it should have s elements of dimensions
     [n x m].
- if samp is an array, then it should be of dimensions [n x m x s].}

\item{Y}{[s] the class labels.}

\item{p}{[n x m x c] the probability per edge of being connected per class for c classes.}

\item{es}{[z] the number of edges to look for, arbitrarily breaking ties as necessary.}

\item{coherent=FALSE}{a logical indicating whether to approximate a coherent, or incoherent, subgraph.}

\item{tstat="fisher"}{the test statistic to use. options are fisher's exact ("fisher") and chi-squared ("chisq").}
}
\value{
subgraph [n x m] an array indicating whether an edge is present or not present in the subgraph.

p [n x m x c] the probability per edge of being connected per class for c classes.
}
\description{
\code{sg.bern.compute_graph_statistics} estimates the graph statistics necessary to compute a subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
