% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sg.beta.estimator}
\alias{sg.beta.estimator}
\title{A function to fit a beta distribution to a given univariate samp of data.}
\usage{
sg.beta.estimator(samp)
}
\arguments{
\item{samp}{a univariate samp of values with arbitrary class labels.}
}
\value{
alpha [n x m] the alpha parameter per edge.

beta [n x m] the beta parameter per edge.
}
\description{
\code{sg.beta.estimator} uses the method of moments to estimate the parameters of a beta
distribution, alpha and beta, for a given samp.
}
\seealso{
\code{\link{sg.beta.graph_estimator}}
}
