% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_estimator.R
\name{sg.beta.graph_estimator}
\alias{sg.beta.graph_estimator}
\title{A function to fit a beta distribution to the edges of the matrices in a graph.}
\usage{
sg.beta.graph_estimator(samp)
}
\arguments{
\item{samp}{a list or array of graphs with arbitrary labelling.
 - if samp is a list, then it should have s elements of dimensions
     [n x m].
- if samp is an array, then it should be of dimensions [n x m x s].}
}
\value{
params$alpha the alpha parameter per edge. [n x m]

params$beta the beta parameter per edge. [n x m]
}
\description{
\code{sg.beta.graph_estimator} uses the method of moments to estimate the parameters of a beta
distribution, alpha and beta, for a collection of graphs.
}
\seealso{
\code{\link{sg.beta.estimator}}
}
