% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sg.beta.sample_graph}
\alias{sg.beta.sample_graph}
\title{Sample Graphs from Beta Distribution}
\usage{
sg.beta.sample_graph(alpha, beta, s = 10, type = "array")
}
\arguments{
\item{alpha}{a [n x n] matrix indicating the alphas of each edge.}

\item{beta}{a [n x n] matrix indicating the betas of each edge.}

\item{s}{the number of graphs to sample.}

\item{type}{an option of "list" or "array" (default) for the format of the output.}
}
\value{
sample the graph observations from the given beta distribution.
    - if type == "list", returns a p element list of [n x n] observations.
    - if type == "array" (default), returns a [n x n x s] element array where the 3rd dimension indexes the observations.
}
\description{
\code{sg.beta.sample_graph} A function to sample from a graph-valued RV where each edge follows a beta distribution.
}
\seealso{
\code{\link{list2array}} \code{\link{array2list}}
}
