% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xval_data_split.R
\name{sg.xval_split_data}
\alias{sg.xval_split_data}
\title{Cross-Validation Data Splitter}
\usage{
sg.xval_split_data(samp, Y, holdout)
}
\arguments{
\item{samp}{[n x n x s] an array of input data.}

\item{Y}{[s] the class labels.}

\item{holdout}{[z] the examples to hold out for the validation set.}
}
\value{
train_set [n x n x (s - z)] the training examples.

train_y [(s - z)] the labels for the training set.

test_set [n x n x s] the testing examples.

test_y [s] the labels for the testing set.
}
\description{
\code{sg.bern.xval_split_data} Bernoulli Subgraph Classifier with Cross Validation to determine the optimal subgraph.
}
