% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph_classifier.R
\name{sg.bern.subgraph_classifier}
\alias{sg.bern.subgraph_classifier}
\title{Bernoulli Subgraph Classifier}
\usage{
sg.bern.subgraph_classifier(test, test_results, p, pi, e, coherent = FALSE)
}
\arguments{
\item{test}{the test array to generate labels for.
 - if test is a list, then it should have s elements of dimensions
     [n x m].
- if test is an array, then it should be of dimensions [n x m x s].}

\item{p}{[n x m x c] the probability per edge of being connected per class for c classes.}

\item{pi}{[c] the pi vector probability of each class occuring.}

\item{e}{the number of edges to look for, arbitrarily breaking ties as necessary.}

\item{test_stats}{[n x m] the test-statistic results.}

\item{coherent=FALSE}{a logical indicating whether to approximate a coherent, or incoherent, subgraph.}
}
\value{
edges [e] the edges present in the subgraph.

Yhat [s] the predicted class for each test example.
}
\description{
\code{sg.bern.subgraph_classifier} classifies arbitrary data given the estimators of a subgraph.
}
\seealso{
\code{\link{sg.bern.compute_graph_statistics}}
}
